requirement.txt
django
djangorestframework
drf-yasg
django-cors-headers
djangorestframework-simplejwt